Kalender v1.11.5 (#1252)
* Remove no longer needed "ilch-date" CSS class.
* Code style improvements.
* Some improvements to the code.
* Fix naming of model and some variables.

---------

+++ application/modules/calendar/controllers/admin/Index.php
@@ -165,7 +165,7 @@ class Index extends \Ilch\Controller\Admin
 
     public function delAction()
     {
-        if ($this->getRequest()->isSecure()) {
+        if ($this->getRequest()->isSecure() && !empty($this->getRequest()->getParam('id'))) {
             $calendarMapper = new CalendarMapper();
             $calendarMapper->delete($this->getRequest()->getParam('id'));
